tracker/easy: avoid race & crash on exit
cv::destroyWindow() can't be called along with `cv::imshow()' in
progress from another thread. Moreover, `cv::waitKey()' doesn't work
without any windows present.

Also check for preview enablement to avoid a crash in
`cv::destroyWindow()'.

+++ tracker-easy/tracker-easy.cpp
@@ -88,10 +88,11 @@ namespace EasyTracker
 
     Tracker::~Tracker()
     {
-        cv::destroyWindow("Preview");
-
         iThread.exit();
         iThread.wait();        
+
+        if (iDebug)
+            cv::destroyWindow("Preview");
         
         if (camera)
         {
